refactor(app-hero): refactor app-hero and app-nav-header
found issue with contact page background, will bring up to elizabeth tomorrow
