+++ src/components/app-hero-home/app-hero-home.scss
@@ -1,7 +1,7 @@
 // hero
 .hero-home {
   align-items: flex-end;
-  background-color: $black;
+  // background-color: $black;
   background-position: top center;
   background-size: cover;
   display: flex;
