+++ src/components/app-hero/app-hero.scss
@@ -1,6 +1,6 @@
 // hero
 .hero {
-  background-color: $black;
+  // background-color: $black;
   background-position: top center;
   background-size: cover;
   height: 100vh;
