+++ src/components/app-hero/app-hero.tsx
@@ -1,4 +1,4 @@
-import { Component, Prop, Element } from '@stencil/core';
+import { Component, Prop } from '@stencil/core';
 
 @Component({
   tag: 'app-hero',
@@ -9,17 +9,20 @@ export class AppHero {
   @Prop() backgroundUrl: string;
   @Prop() textNoWrap: boolean;
   @Prop() hideLink: boolean;
-  @Element() host: Element;
 
   componentDidLoad() {
     let element;
     try {
-      console.log('started app-hero didLoad');
-      element = document.querySelector('header.hero') as HTMLElement;
-      console.log('completed app-hero didLoad');
+      element = document.querySelector('header.hero');
     } catch (e) {
       console.log('completed app-hero undefined', e);
     }
+    console.log(
+      'Element = ',
+      element,
+      ' with background ',
+      `url('${this.backgroundUrl}')`
+    );
     element.style.backgroundImage = `url('${this.backgroundUrl}')`;
   }
