+++ src/pages/open-forge-app/open-forge-app.scss
@@ -9,10 +9,10 @@ nav {
     box-shadow: 0 0 6px 0 rgba(0, 0, 0, 0.3);
   }
 
-  &.background {
-    background: $black;
-    box-shadow: 0 0 6px 0 rgba(0, 0, 0, 0.3);
-  }
+  // &.background {
+  //   background: $black;
+  //   box-shadow: 0 0 6px 0 rgba(0, 0, 0, 0.3);
+  // }
 }
 
 .navbar-brand {
